% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_a5064}
\alias{county_a5064}
\title{County-Level Estimates of Health Insurance Coverage for Age 50-64}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28283 rows and 14 columns.}
\usage{
county_a5064
}
\description{
Available years: 2008-2016. Estimates are specific to those age 50-64,
all race, both sexes, and for all income levels.
}
\keyword{datasets}
