% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_yr0816_ag0018}
\alias{county_yr0816_ag0018}
\title{County-Level Estimates of Health Insurance Coverage: Age 00-18}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28283 rows and 16 columns.}
\usage{
county_yr0816_ag0018
}
\description{
Available years: 2008-2016. Estimates are specific to those age 00-18,
are of any race, any sex, and for all income levels.
}
\keyword{datasets}
