% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_yr1016_ag5064}
\alias{county_yr1016_ag5064}
\title{County-Level Estimates of Health Insurance Coverage for Age 50-64}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21998 rows and 16 columns.}
\usage{
county_yr1016_ag5064
}
\description{
Available years: 2010-2016. Estimates are specific to those age 50-64,
are of any race, any sex, and for all income levels.
}
\keyword{datasets}
