% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sahie_county}
\alias{sahie_county}
\title{County-Level Number of Uninsured Eestimates}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28281 rows and 5 columns.}
\usage{
sahie_county
}
\description{
Use \code{sahie_county}.
}
\keyword{datasets}
